Remove the Gemfile.lock from the gem files
The Gemfile.lock was removed in #11 but it was missed to be removed from the files being included in the released gem. This is breaking our integration tests and it is still attempting to publish the removed file as part of the gem.

`make pr` did not fail as it runs `bundler init` before running the unit and smoke tests.

On a clean branch, running `make integ-tests` reproduces the failure.